requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
packaging==20.1
pyparsing==2.4.6
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
